Bit Rework: use get token rather than assume role
This gives better auditing since the user remains the same.
To do this, introduce polcies that deny access if not called with MFA
and https.

Also swap handling of aws credentials from a host var to a group var
for consistency with the pdf-signer project.

Also expand the permissions granted to the delegate admin to include
enough for pdf-signer automation to use a delegate admin.